/src/com-port-api reactor::start method introduced
Since the reactor is now an abstract class, it is necessary to give
subclasses a chance to properly start the working thread by passing to
its constructor a right `this` object.

The new method must be called by the user in order to start the
reactor.

The previous `start` function which was passed to the worker thread
constructor renamed to `run`.

Example file is also changed to reflect API changes.